refactor(engine): Separate expression functions into its own module
