+++ tracecat/expressions/engine.py
@@ -42,11 +42,9 @@ For secrets, you should:
 from __future__ import annotations
 
 import json
-import operator
 import re
 from collections.abc import Iterable, Iterator
 from dataclasses import dataclass
-from datetime import datetime
 from enum import IntEnum, StrEnum, auto
 from typing import Any, Generic, TypeVar
 
@@ -54,80 +52,15 @@ import jsonpath_ng
 from jsonpath_ng.exceptions import JsonPathParserError
 
 from tracecat.expressions import patterns
+from tracecat.expressions.functions import BUILTIN_TYPE_NAPPING, FUNCTION_MAPPING
 from tracecat.types.exceptions import TracecatExpressionError
 
 T = TypeVar("T")
 
-
-def _bool(x: Any) -> bool:
-    if isinstance(x, bool):
-        return x
-    if isinstance(x, str):
-        return x.lower() in ("true", "1")
-    # Use default bool for everything else
-    return bool(x)
-
-
-def _from_timestamp(x: int, unit: str) -> datetime:
-    if unit == "ms":
-        dt = datetime.fromtimestamp(x / 1000)
-    else:
-        dt = datetime.fromtimestamp(x)
-    return dt
-
-
-_BUILTIN_TYPE_NAP = {
-    "int": int,
-    "float": float,
-    "str": str,
-    "bool": _bool,
-    # TODO: Perhaps support for URLs for files?
-}
 ExprStr = str
 """An annotated template string that can be resolved into a type T."""
 
 OperandType = dict[str, Any]
-# Supported Formulas / Functions
-_FN_TABLE = {
-    # Comparison
-    "less_than": operator.lt,
-    "less_than_or_equal": operator.le,
-    "greater_than": operator.gt,
-    "greater_than_or_equal": operator.ge,
-    "not_equal": operator.ne,
-    "is_equal": operator.eq,
-    "not_null": lambda x: x is not None,
-    "is_null": lambda x: x is None,
-    # Regex
-    "regex_extract": lambda pattern, text: re.search(pattern, text).group(0),
-    "regex_match": lambda pattern, text: bool(re.match(pattern, text)),
-    "regex_not_match": lambda pattern, text: not bool(re.match(pattern, text)),
-    # Collections
-    "contains": lambda item, container: item in container,
-    "does_not_contain": lambda item, container: item not in container,
-    "length": len,
-    "is_empty": lambda x: len(x) == 0,
-    "not_empty": lambda x: len(x) > 0,
-    # Math
-    "add": operator.add,
-    "sub": operator.sub,
-    "mul": operator.mul,
-    "div": operator.truediv,
-    "mod": operator.mod,
-    "pow": operator.pow,
-    # Transform
-    "join": lambda items, sep: sep.join(items),
-    "concat": lambda *items: "".join(items),
-    # Logical
-    "and": lambda a, b: a and b,
-    "or": lambda a, b: a or b,
-    "not": lambda a: not a,
-    # Type conversion
-    # Convert JSON to string
-    "serialize_json": json.dumps,
-    # Convert timestamp to datetime
-    "from_timestamp": lambda x, unit,: _from_timestamp(x, unit),
-}
 
 
 class Expression:
@@ -375,7 +308,7 @@ class ExpressionParser:
         return "${{ " + expr + " }}"
 
     def _cast_result(self, result: Any, rtype: str):
-        return _BUILTIN_TYPE_NAP[rtype](result)
+        return BUILTIN_TYPE_NAPPING[rtype](result)
 
     def _maybe_parse_jsonpath(
         self, expr: str, expr_context: ExprContext, depth: int = 0
@@ -406,7 +339,7 @@ class ExpressionParser:
         if ExprContext.FN in self._exclude_contexts:
             return expr
         resolved_args = self._parse_parameter_pack(args, depth + 1)
-        result = _FN_TABLE[qualname](*resolved_args)
+        result = FUNCTION_MAPPING[qualname](*resolved_args)
         return result
 
     def _parse_parameter_pack(self, expr: str, depth: int = 0) -> tuple:
@@ -428,8 +361,6 @@ class ExpressionParser:
             raise ValueError(
                 f"Invalid iterator collection: {iter_collection_expr!r}. Must be an iterable."
             )
-
-        print(f"Collection: {collection}")
 
         # Reset the loop flag
         self._flags &= ~ParserFlags.LOOP
